........
  r25980 | sandra.castro | 2013-08-09 03:09:40 -0600 (Fri, 09 Aug 2013) | 13 lines
  
  
         New Development: No
              JIRA Issue: CAS-5485
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: test_flagdata
    Put in Release Notes: No
               Module(s): flagging
             Description: Fixed a bug that prevented manually flagging an antenna in list mode if autocorr=True was also present in another command in the list. A new test simulating this problem was added (test_autocorr_wvr_list).
  
  	
........
  r25983 | sanjay.bhatnagar | 2013-08-09 09:42:10 -0600 (Fri, 09 Aug 2013) | 39 lines
  
  
         New Development: Yes
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: Yes
               Module(s): Module Names change impacts.
             Description: Describe your changes here...
  
  Utils.cc: Modifed nearestNdx calculation to a more efficient
                algo. 
  
  CFBuffer.{cc,h}: Added the asStruct() and related methods. 
  
  VLACalcIlluminationConvFunc.{cc,h}: Changes to compute only the
                                      required CF pol. planes.
  
  BeamCalc.{cc,h}: Changes to compute only the required CF pol. planes. 
  
  AWProjectFT.{cc,h}: Added CFBStruct data member and related code
                      changes. 
  
  CFCell.{cc,h}: Added asStruct() and related methods. 
  
  VisibilityResamplerBase.h: Added cacheAxisInc(Int *)
  
  AWVisResampler.cc: Using sincos
  
  EVLAAperture.h, AWProjectWBFT.cc, ATerm.h: Minor changes
  
  WTerm.cc: Using sincos
  
  VBStore.h: Added CFBStruct data member
  
  VLAIlluminationConvFunc.cc: Using SynthesisUtils::makeUVCoords().
........
  r25986 | sanjay.bhatnagar | 2013-08-09 14:39:18 -0600 (Fri, 09 Aug 2013) | 14 lines
  
  
         New Development: Yes
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: Yes
               Module(s): Module Names change impacts.
             Description: Describe your changes here...
  
  Test code for ARD.  Nothing depends on it yet.	
........
  r25989 | urvashi.rao | 2013-08-09 16:42:52 -0600 (Fri, 09 Aug 2013) | 14 lines
  
  
         New Development: Yes
              JIRA Issue: Imager Refactor
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: --
           Test Programs: List test programs
    Put in Release Notes: No
               Module(s): Module Names change impacts.
             Description: Removed the list of SI tools from PySynthesisImager and modified selection par dict accordingly. Added a ParallelCubeClean class with the same interface as all the others, so they can be called using the same task wrappers ( they = Serial runs, Parallel Major Cycle only, Parallel Major/Minor cycles for Cube imaging ).  Removed some log messages.
  
  	
........
  r25991 | george.moellenbrock | 2013-08-09 17:30:04 -0600 (Fri, 09 Aug 2013) | 17 lines
  
  
         New Development: No
              JIRA Issue: CAS-3517,4621
          Ready for Test: Yes
       Interface Changes: Yes
  What Interface Changed: New calibrater_cmpt methods, mainly
           Test Programs: tCLPatchPanel
    Put in Release Notes: Yes
               Module(s): synthesis, calibrater tool
             Description: Initial check-in of cal library c++, mainly to
  			ensure happy co-existence with existing
  			calibration apply framework.  Additional 
  			callib deployment (in tasks, etc., w/ 
  			unit tests and regresions) coming soon.
........
  r25992 | george.moellenbrock | 2013-08-09 17:38:26 -0600 (Fri, 09 Aug 2013) | 13 lines
  
  
         New Development: No
              JIRA Issue: CAS-3517,4621
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: tCLPatchPanel
    Put in Release Notes: No
               Module(s): code/synthesis
             Description: This is the test program omitted from r25911.
........
  r26014 | michel.caillat | 2013-08-12 10:33:39 -0600 (Mon, 12 Aug 2013) | 14 lines
  
  
         New Development: No
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: 
    Put in Release Notes: No
               Module(s): alma
             Description: Fixed a bug reported by Dirk Petry when the filler was used in lazy mode and was presented pure auto data. Also committed the code to process the Ephemeris table.
  
  	
........
  r26029 | urvashi.rao | 2013-08-12 18:51:20 -0600 (Mon, 12 Aug 2013) | 15 lines
  
  
  	
        New Development: Yes
              JIRA Issue: Imager Refactor
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: No
               Module(s): Module Names change impacts.
             Description: Started adding the true Hogbom fortran call to the new Minor cycle hogbom class. Renamed the simple hogbom to algo='test'. Updated test program to use 'test' only for now.
  
  	
........


git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/test@26094 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/synthesis/CalTables/NewCalTable.h
@@ -214,8 +214,8 @@ class NewCalTable : public Table
    // Provide RO access to 
    //   inline ROCTMainColumns mcols() const {return ROCTMainColumns(*this);};
 
-   static Complex NCTtestvalueC(Int iant,Int ispw,Int ich,Double time,Double refTime,Double tint);
-   static Float NCTtestvalueF(Int iant,Int ispw,Int ich,Double time,Double refTime,Double tint);
+   static Complex NCTtestvalueC(Int iant,Int ispw,Double ich,Double time,Double refTime,Double tint);
+   static Float NCTtestvalueF(Int iant,Int ispw,Double ich,Double time,Double refTime,Double tint);
 
    // Fill in antenna-based solutions
    //  (obsId-less version calls obsId-capable version w/ obsId=0)
